+++ Source/Core/AST/Program.cs
@@ -498,7 +498,7 @@ public class Program : Absy
     return callGraph;
   }
 
-  public static Graph<Block> GraphFromBlocks(List<Block> blocks, bool forward = true)
+  public static Graph<Block> GraphFromBlocks(IReadOnlyList<Block> blocks, bool forward = true)
   {
     var result = new Graph<Block>();
     if (!blocks.Any())
